Refactor CAN motor control examples
